requirements.txt
python-telegram-bot==20.7
google-generativeai==0.8.3
chromadb==0.5.23
sentence-transformers==3.3.1
python-dotenv==1.0.1
